Decouple mode base cluster part 1
