+++ src/app/clusters/mode-base-server/app_config_dependent_sources.cmake
@@ -16,8 +16,10 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
+    "${CLUSTER_DIR}/CodegenIntegration.h"
+    "${CLUSTER_DIR}/Delegate.h"
     "${CLUSTER_DIR}/mode-base-cluster-objects.cpp"
     "${CLUSTER_DIR}/mode-base-cluster-objects.h"
-    "${CLUSTER_DIR}/mode-base-server.cpp"
     "${CLUSTER_DIR}/mode-base-server.h"
+    "${CLUSTER_DIR}/ModeBaseCluster.cpp"
 )
