+++ src/app/clusters/mode-base-server/app_config_dependent_sources.gni
@@ -12,8 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
+  "CodegenIntegration.h",
+  "Delegate.h",
   "mode-base-cluster-objects.cpp",
   "mode-base-cluster-objects.h",
-  "mode-base-server.cpp",
   "mode-base-server.h",
+  "ModeBaseCluster.cpp",
 ]
